server/requirements.txt
# Sistema de Monitoramento IoT para Ônibus
# Dependências unificadas para produção e desenvolvimento

# Framework web
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Banco de dados
psycopg2-binary==2.9.7

# Processamento de imagens e visão computacional
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

# Machine Learning e YOLO (para análise de ocupação)
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2

# Análise de dados
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2

# Análise de tempo e previsões
prophet==1.1.4

# Requisições HTTP
requests==2.31.0

# Utilitários
python-dotenv==1.0.0
python-dateutil==2.8.2

# Desenvolvimento e testes (opcional)
pytest==7.4.2
pytest-flask==1.2.0
